x64 warnings in DialogBuilder.

+++ far/DlgBuilder.hpp
@@ -181,7 +181,7 @@ class DialogBuilderBase
 		int ColumnBreakIndex;
 		int ColumnStartY;
 		int ColumnEndY;
-		int ColumnMinWidth;
+		intptr_t ColumnMinWidth;
 
 		static const int SECOND_COLUMN = -2;
 
@@ -303,7 +303,7 @@ class DialogBuilderBase
 				if (MaxWidth < Width)
 					MaxWidth = Width;
 			}
-			int ColumnsWidth = 2*ColumnMinWidth+1;
+			intptr_t ColumnsWidth = 2*ColumnMinWidth+1;
 			if (MaxWidth < ColumnsWidth)
 				return ColumnsWidth;
 			return MaxWidth;
@@ -478,7 +478,7 @@ class DialogBuilderBase
 			Item->X1 = 5 + Indent;
 			Item->X2 = Item->X1 + ItemWidth(*Item) - 1;
 
-			int RelativeToWidth = RelativeTo->X2 - RelativeTo->X1;
+			intptr_t RelativeToWidth = RelativeTo->X2 - RelativeTo->X1;
 			RelativeTo->X1 = Item->X2 + 2;
 			RelativeTo->X2 = RelativeTo->X1 + RelativeToWidth;
 
@@ -542,7 +542,7 @@ class DialogBuilderBase
 		{
 			for(int i=ColumnStartIndex; i<DialogItemsCount; i++)
 			{
-				int Width = ItemWidth(DialogItems [i]);
+				intptr_t Width = ItemWidth(DialogItems [i]);
 				if (Width > ColumnMinWidth)
 					ColumnMinWidth = Width;
 				if (i >= ColumnBreakIndex)
